Add warning text about curl install method
Installing software by piping random scripts into a shell is a dangerous
way to install software. Although it is convenient to have a simple
install procedure like that, it is highly discouraged in most tech
circles and I would not like to perpetuate the cycle. Adding the warning
disclaimer allows me to continue offering the curl install procedure
while educating users that it is an unsafe way to install software and
lets me point them towards safe install procedures.