Clarify fishmarks update instructions
The update instructions did not account for users whose fishmarks
install locations were not the default ~/.fishmarks, but the new
instructions note that ~/.fishmarks is just the default and include code
demonstrating how to change into the install directory. The instructions
for step 2 and 3 explain the function of each of the Git commands.