Update valid name to regex to match entire string
The regular expression used for verifying that bookmark names in the
correct format should match the entire string being passed to it. With
out the characters for matching the beginning and end of the string, the
regular expression would accept almost all input the regular expression
would match a small portion of the string and then automatically return
true.